Django-rfw/zuri2/requirement.txt
python
Django==4.2.4
djangorestframework==3.14.0
